Singleton and Huge Refactoring
