+++ src/components/Tray/index.tsx
@@ -1,11 +1,10 @@
-import AsyncStorage from "@react-native-async-storage/async-storage"
 import { useNavigationState } from "@react-navigation/native"
 import React, { FC, useEffect } from "react"
-import { Dimensions, Pressable, View } from "react-native"
+import { Dimensions, View } from "react-native"
 import { newsSheetEventEmitter } from "utils/events"
-import { useNotificationBadge } from "utils/notifications"
 import { LittleTitle } from "./LittleTitle"
 import { TrayButton } from "./TrayButton"
+import { useNotificationBadge } from "notifications/useNotificationBadge"
 
 /**
  * Top bar menu present in every page of the application, containing the
@@ -49,50 +48,7 @@ export const Tray: FC<{
       }}
     >
       <LittleTitle titleInCorner={notInHome || newsOpen} />
-      {/* <Pressable
-        style={{
-          width: 20,
-          height: 20,
-          backgroundColor: "yellow",
-          marginRight: 15,
-        }}
-        onPress={async () => {
-          const notif = await getAllNotificationsFromStorage()
-          console.log(notif[0].notification.content.data.content)
-          console.log(notif)
-        }}
-      /> */}
-      <Pressable
-        style={{
-          width: 20,
-          height: 20,
-          backgroundColor: "green",
-          marginRight: 15,
-        }}
-        onPress={() => {
-          void AsyncStorage.setItem("notifications", JSON.stringify([]))
-        }}
-      />
-      {/* 
-      
-      <Pressable
-        style={{
-          width: 20,
-          height: 20,
-          backgroundColor: "red",
-          marginRight: 15,
-        }}
-        onPress={async () => {
-          console.log("logging current permission: ")
-          await notificationsTestingUtils.logPermission()
 
-          const granted = await notificationsTestingUtils.askPermission()
-          console.log("granted: " + granted)
-
-          console.log("logging new permission: ")
-          await notificationsTestingUtils.logPermission()
-        }}
-      /> */}
       <TrayButton label="downloads" onClick={() => props.onDownloads()} />
       <TrayButton
         label="notifications"
