+++ tsconfig.json
@@ -12,7 +12,9 @@
       "utils/*": ["src/utils/*"],
       "contexts/*": ["src/contexts/*"],
       "api": ["src/api"],
-      "api/*": ["src/api/*"]
+      "api/*": ["src/api/*"],
+      "notifications/*": ["src/notifications/*"]
+
     },
     "forceConsistentCasingInFileNames": true
   }
